pyproject.toml
[project]
name = "mcp-a2a"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.0",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "cos-python-sdk-v5>=1.9.41",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.25.0",
    "openai>=2.15.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.45",
    "timeloop>=1.0.2",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
